requirements.txt
# requirements.txt
openai>=0.27.0
requests>=2.28.0
jinja2>=3.1.0
python-dotenv>=1.0.0
pandas>=1.5.0
retrying>=1.3.3
pyyaml>=6.0.0
click>=8.1.0
colorama>=0.4.6
loguru>=0.7.0
